requirements.txt
Flask==3.1.1
matplotlib==3.7.5
numpy==1.26.4  # 修改为与 matplotlib 3.7.5 兼容的版本
Pillow==11.2.1
torch==2.7.0
torchvision==0.22.0
Werkzeug==3.1.3
gunicorn==21.2.0